requirements.txt
paho-mqtt==1.3.1
ruamel.yaml==0.15.34
python-jenkins==0.4.15
hypchat==0.21
